Refactored user bookings with booking card compontent and past stays section
